requirements.txt
pycryptodome==3.18.0
Requests==2.31.0
urllib3==1.26.16
urllib3==1.26.5
